Add the initial codebase for the bundle
The code was based on a Openconext-profile branch where the basis of
this bundle where built in a non-reusable bundle setup.

The reusable bundle 'skeleton' has been taken from the Stepup-bundle

The Ibuildings qa tools dependency was dropped in favor of installing
qa related tools manually, anticipating the deprication of IB QA tools.